% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Country Population Over Time Graph}
\usage{
CountryPopulation(countryName)
}
\arguments{
\item{countryName}{The name of a country in the UN}
}
\description{
This function takes a country in the United Nations and spits out a graph
of that countries population over time.
}
\examples{
CountryPopulation('Brazil')
CountryPopulation('France')
}
