% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulationDesc.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Population Over Time in UN Countries}
\format{
a data frame with 16685 observations
\describe{
  \item{Country}{The name of a country in the UN.}
  \item{Year}{The year.}
  \item{Population}{The population for that year.}
}
}
\source{
{Dr. Luedeker}
}
\usage{
WorldPopulation
}
\description{
A dataset of all the countries in the United Nations and their population from 1950 to 2020.
}
\keyword{datasets}
